Apply Clippy suggestions

+++ rand_core/src/block.rs
@@ -381,13 +381,13 @@ impl<R: BlockRngCore<Item = u64>> RngCore for BlockRng64<R> {
         let mut read_len = 0;
         self.half_used = false;
         while read_len < dest.len() {
-            if self.index as usize >= self.results.as_ref().len() {
+            if self.index >= self.results.as_ref().len() {
                 self.core.generate(&mut self.results);
                 self.index = 0;
             }
 
             let (consumed_u64, filled_u8) = fill_via_u64_chunks(
-                &self.results.as_ref()[self.index as usize..],
+                &self.results.as_ref()[self.index..],
                 &mut dest[read_len..],
             );
